Move dimension conversions to separate file
